Moving find_or_create_qualification
