+++ docs/web/docs/guides/how_to_use/worker_quality/common_qualification_flows.md
@@ -75,7 +75,7 @@ shared_state.qualifications = [
 ### Adding custom qualifications to SharedTaskState
 You should be able to specify a qualification in Mephisto using the following:
 ```python
-from mephisto.utils.misc import find_or_create_qualification
+from mephisto.data_model.qualification import find_or_create_qualification
 
 find_or_create_qualification(db, "MY_QUALIFICATION_NAME")
 ```
