+++ mephisto/abstractions/blueprints/mixins/onboarding_required.py
@@ -16,8 +16,11 @@ from typing import (
 from mephisto.abstractions.blueprint import BlueprintMixin
 from dataclasses import dataclass, field
 from omegaconf import MISSING, DictConfig
-from mephisto.data_model.qualification import make_qualification_dict, QUAL_NOT_EXIST
-from mephisto.utils.misc import find_or_create_qualification
+from mephisto.data_model.qualification import (
+    make_qualification_dict,
+    QUAL_NOT_EXIST,
+    find_or_create_qualification,
+)
 
 if TYPE_CHECKING:
     from mephisto.abstractions.blueprint import SharedTaskState
