+++ mephisto/abstractions/blueprints/mixins/screen_task_required.py
@@ -21,8 +21,11 @@ import types
 from mephisto.abstractions.blueprint import BlueprintMixin
 from dataclasses import dataclass, field
 from omegaconf import MISSING, DictConfig
-from mephisto.data_model.qualification import make_qualification_dict, QUAL_NOT_EXIST
-from mephisto.utils.misc import find_or_create_qualification
+from mephisto.data_model.qualification import (
+    make_qualification_dict,
+    QUAL_NOT_EXIST,
+    find_or_create_qualification,
+)
 
 
 if TYPE_CHECKING:
